pyproject.toml
[project]
name = "rag-using-airbnb-reviews-to-augment-touristic-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "numpy>=2.3.4",
    "pyodbc>=5.3.0",
    "requests>=2.32.5",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.0",
]
